requirements.txt
SQLAlchemy==2.0.32
pyTelegramBotAPI==4.12.0
requests==2.28.2
geopy==2.4.1
python-dotenv
psycopg2
uvicorn
fastapi